+++ crates/aranya-metrics/Cargo.toml
@@ -28,7 +28,7 @@ metrics = { workspace = true }
 metrics-exporter-prometheus = { version = "0.17", features = ["uds-listener"] }
 metrics-util = { workspace = true }
 serde = { workspace = true }
-sysinfo = { workspace = true, features = ["serde"] }
+sysinfo = { workspace = true }
 tempfile = { workspace = true }
 tokio = { workspace = true, features = ["process"] }
 tracing = { workspace = true }
